+++ app.go
@@ -7,10 +7,9 @@ import (
 	"github.com/bitrise-io/appcenter/model"
 )
 
-// App ...
-type App struct {
-	client          Client
-	owner, name     string
+// AppAPI ...
+type AppAPI struct {
+	api             API
 	commandExecutor commander.CommandExecutor
 }
 
@@ -20,45 +19,27 @@ type App struct {
 // 2) Fetches the releases and gets the latest because it is the recent uploaded release.
 // 3) Fetches the lastest release full data.
 // 4) Sets the remaining groups on the release with the API. Because AppCenter CLI is not able to set the groups in one command.
-func (a App) NewRelease(opts ReleaseOptions) (Release, error) {
+func (a AppAPI) NewRelease(opts model.ReleaseOptions) (model.Release, error) {
 	//upload the artifact with the AappCenter CLI
 	commandArgs := a.createCLICommandArgs(opts)
 	str, err := a.commandExecutor.ExecuteCommand("appcenter", commandArgs...)
 	if err != nil {
-		return Release{}, fmt.Errorf("Failed to create AppCenter release: %s", str)
+		return model.Release{}, fmt.Errorf("Failed to create AppCenter release: %s", str)
 	}
 
 	fmt.Println(fmt.Sprintf("Command execution result: %s", str))
 
 	//fetch releases and find the latest
-	api := API{Client: a.client}
-	release, err := api.GetLatestReleases(a)
+	release, err := a.api.GetLatestReleases(opts.App)
 	if err != nil {
-		return Release{}, err
-	}
-
-	release.app = a
-
-	// set the groups on the app
-	if len(opts.GroupNames) > 1 {
-		for _, groupName := range opts.GroupNames[1:] {
-			if len(groupName) == 0 {
-				continue
-			}
-			group, err := a.Groups(groupName)
-			if err != nil {
-				return Release{}, err
-			}
-
-			release.AddGroup(group, opts.Mandatory, opts.NotifyTesters)
-		}
+		return model.Release{}, err
 	}
 
 	return release, nil
 }
 
-func (a App) createCLICommandArgs(opts ReleaseOptions) []string {
-	appName := a.owner + "/" + a.name
+func (a AppAPI) createCLICommandArgs(opts model.ReleaseOptions) []string {
+	appName := opts.App.Owner + "/" + opts.App.AppName
 	commandArgs := []string{"distribute", "release", "-a", appName, "-f", opts.FilePath, "-g", opts.GroupNames[0]}
 
 	if len(opts.BuildNumber) != 0 {
@@ -75,15 +56,11 @@ func (a App) createCLICommandArgs(opts ReleaseOptions) []string {
 }
 
 // Groups ...
-func (a App) Groups(name string) (model.Group, error) {
-	api := API{Client: a.client}
-
-	return api.GetGroupByName(name, a)
+func (a AppAPI) Groups(name string, app model.App) (model.Group, error) {
+	return a.api.GetGroupByName(name, app)
 }
 
 // Stores ...
-func (a App) Stores(name string) (model.Store, error) {
-	api := API{Client: a.client}
-
-	return api.GetStore(name, a)
+func (a AppAPI) Stores(name string, app model.App) (model.Store, error) {
+	return a.api.GetStore(name, app)
 }
