+++ model/app.go
@@ -1,0 +1,6 @@
+package model
+
+// App ...
+type App struct {
+	Owner, AppName string
+}
