+++ model/release_options.go
@@ -1,0 +1,12 @@
+package model
+
+// ReleaseOptions ...
+type ReleaseOptions struct {
+	BuildVersion  string
+	BuildNumber   string
+	GroupNames    []string
+	Mandatory     bool
+	NotifyTesters bool
+	FilePath      string
+	App           App
+}
